utilities: ram_console: record a absolute position
Use a absolute position number but not a postion in RAM buffer, so it
can make it easier to only display new log if using continues dumpping.

+++ utilities/debug_console/fsl_ram_console.h
@@ -15,7 +15,7 @@ struct ram_console_header {
     char flag_string[12];
     char *ram_console_buf_addr;
     size_t ram_console_buf_size;
-    size_t pos;
+    uint32_t pos;
 };
 
 #endif /* __FSL_RAM_CONSOLE_H__ */
